[#315] refactor: dto 패키지 application 패키지 하위로 이동
