+++ src/main/java/com/beat/domain/member/application/dto/AccessTokenGenerateResponse.java
@@ -1,4 +1,4 @@
-package com.beat.domain.member.dto;
+package com.beat.domain.member.application.dto;
 
 public record AccessTokenGenerateResponse(
 	String accessToken
