+++ src/main/java/com/beat/domain/member/application/dto/LoginSuccessResponse.java
@@ -1,4 +1,4 @@
-package com.beat.domain.member.dto;
+package com.beat.domain.member.application.dto;
 
 public record LoginSuccessResponse(
 	String accessToken,
